backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
supabase>=2.0.1
pydantic[email]>=2.0.0
gspread>=5.10.0
google-auth>=2.22.0
python-dotenv>=1.0.0
requests>=2.31.0
